windows_bot/requirements.txt
python-dotenv
requests
twitchio
